+++ __brick__/{{project_name.snakeCase()}}/lib/features/home/domain/abstractions/controllers/home_controller.interface.dart
@@ -1,3 +1,0 @@
-import '../../../../../core/abstractions/controller.interface.dart';
-
-abstract class IHomeController extends IController {}