+++ __brick__/{{project_name.snakeCase()}}/lib/features/home/domain/bindings/home_controller.binding.dart
@@ -1,11 +1,11 @@
-import 'package:flutter/cupertino.dart';
+import 'package:flutter/material.dart';
 
-import '../abstractions/controllers/home_controller.interface.dart';
-import '../../presentation/controller/home.controller.dart';
 import '../../../../core/abstractions/field.interface.dart';
 import '../../../../core/builders/field_validator.builder.dart';
 import '../../../../core/inject.dart';
 import '../../../../core/models/react_field.model.dart';
+import '../../presentation/home.controller.dart';
+import 'home_controller.interface.dart';
 
 class HomeControllerBinding {
   static void inject() {