+++ __brick__/{{project_name.snakeCase()}}/lib/features/home/presentation/home.screen.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 
-class HomeScreen extends StatelessWidget {
+import '../../shared/view_controller.interface.dart';
+import '../domain/bindings/home_controller.interface.dart';
+
+class HomeScreen extends ViewController<IHomeController> {
   const HomeScreen({super.key});
 
   @override